refactor: remove old frontpage components and integrate new navbar structure
