--- app/(frontpage)/test/page.tsx
@@ -1,14 +1,0 @@
-
-import { ModeToggle } from '@/app/(frontpage)/test/components/toggle'
-import React from 'react'
-
-export default function Page() {
-  return (
-    <div className='dark:bg-slate-900 bg-slate-400 flex h-screen w-full'>
-      <div className='flex h-full'>
-
-      {/* <ModeToggle /> */}
-      </div>
-    </div>
-  )
-}
