+++ components/navbar/navbar.tsx
@@ -12,34 +12,11 @@ import {
   PopoverGroup,
   PopoverPanel,
 } from '@headlessui/react'
-import {
-  ArrowPathIcon,
-  Bars3Icon,
-  ChartPieIcon,
-  CursorArrowRaysIcon,
-  FingerPrintIcon,
-  SquaresPlusIcon,
-  XMarkIcon,
-} from '@heroicons/react/24/outline'
-import { ChevronDownIcon, PhoneIcon, PlayCircleIcon } from '@heroicons/react/20/solid'
 import Image from 'next/image'
-import { ModeToggle } from './toggle'
-import SearchBar from '@/components/searchbar'
-import { Link } from 'lucide-react'
+import { ChevronDownIcon, Link } from 'lucide-react'
 import Search from '@/components/ui/search'
-
-const products = [
-  { name: 'Analytics', description: 'Get a better understanding of your traffic', href: '#', icon: ChartPieIcon },
-  { name: 'Engagement', description: 'Speak directly to your customers', href: '#', icon: CursorArrowRaysIcon },
-  { name: 'Security', description: 'Your customers’ data will be safe and secure', href: '#', icon: FingerPrintIcon },
-  { name: 'Integrations', description: 'Connect with third-party tools', href: '#', icon: SquaresPlusIcon },
-  { name: 'Automations', description: 'Build strategic funnels that will convert', href: '#', icon: ArrowPathIcon },
-]
-const callsToAction = [
-  { name: 'Watch demo', href: '#', icon: PlayCircleIcon },
-  { name: 'Contact sales', href: '#', icon: PhoneIcon },
-]
-
+import { products,callsToAction } from '@/const/navbar'
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline'
 export default function Navbar1() {
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false)
