+++ const/navbar.ts
@@ -1,0 +1,124 @@
+import {
+    ArrowPathIcon,
+    ChartPieIcon,
+    CursorArrowRaysIcon,
+    FingerPrintIcon,
+    SquaresPlusIcon,
+  } from '@heroicons/react/24/outline'
+  import { ChevronDownIcon, PhoneIcon, PlayCircleIcon } from '@heroicons/react/20/solid'
+export const products = [
+    { name: 'Analytics', description: 'Get a better understanding of your traffic', href: '#', icon: ChartPieIcon },
+    { name: 'Engagement', description: 'Speak directly to your customers', href: '#', icon: CursorArrowRaysIcon },
+    { name: 'Security', description: 'Your customers’ data will be safe and secure', href: '#', icon: FingerPrintIcon },
+    { name: 'Integrations', description: 'Connect with third-party tools', href: '#', icon: SquaresPlusIcon },
+    { name: 'Automations', description: 'Build strategic funnels that will convert', href: '#', icon: ArrowPathIcon },
+  ]
+ export const callsToAction = [
+    { name: 'Watch demo', href: '#', icon: PlayCircleIcon },
+    { name: 'Contact sales', href: '#', icon: PhoneIcon },
+  ]
+
+export const megaMenuData = [
+    {
+      Title: "Farm Supplies & Equipment",
+      services: [
+        {
+          title: "High-Quality Seeds",
+          slug: "quality-seeds",
+          description: "Browse a variety of certified seeds for optimal crop yield and quality.",
+        },
+        {
+          title: "Agricultural Machinery",
+          slug: "farm-machinery",
+          description: "Find top-notch machinery and equipment to streamline your farming operations.",
+        },
+        {
+          title: "Organic Fertilizers & Pesticides",
+          slug: "fertilizers-pesticides",
+          description: "Access eco-friendly fertilizers and pesticides to promote sustainable farming.",
+        },
+      ],
+    },
+    {
+      Title: "Crop Advisory Services",
+      services: [
+        {
+          title: "Virtual Crop Consultations",
+          slug: "crop-consultations",
+          description: "Connect with agronomists for expert advice on crop management and health.",
+        },
+        {
+          title: "Soil Testing Services",
+          slug: "soil-testing",
+          description: "Get detailed soil analysis and recommendations for optimal crop growth.",
+        },
+        {
+          title: "Weather Insights & Alerts",
+          slug: "weather-insights",
+          description: "Receive accurate weather updates and alerts tailored for farmers.",
+        },
+      ],
+    },
+    {
+      Title: "Livestock & Poultry",
+      services: [
+        {
+          title: "Animal Feed & Supplements",
+          slug: "animal-feed",
+          description: "Purchase premium feed and supplements for healthy livestock and poultry.",
+        },
+        {
+          title: "Online Veterinary Support",
+          slug: "vet-support",
+          description: "Consult experienced veterinarians for livestock health and care advice.",
+        },
+        {
+          title: "Livestock Equipment",
+          slug: "livestock-equipment",
+          description: "Explore tools and equipment designed to make livestock farming efficient.",
+        },
+      ],
+    },
+    {
+      Title: "Agri-Marketplace",
+      services: [
+        {
+          title: "Sell Your Produce",
+          slug: "sell-produce",
+          description: "List your agricultural produce and connect with buyers directly.",
+        },
+        {
+          title: "Farmers' Auctions",
+          slug: "farmers-auctions",
+          description: "Participate in real-time auctions to sell or buy agricultural goods.",
+        },
+        {
+          title: "Market Price Updates",
+          slug: "market-prices",
+          description: "Stay updated on the latest market prices for crops and livestock.",
+        },
+      ],
+    },
+    {
+      Title: "Sustainable Farming",
+      services: [
+        {
+          title: "Organic Farming Resources",
+          slug: "organic-farming",
+          description: "Access tools and techniques for sustainable and organic farming practices.",
+        },
+        {
+          title: "Renewable Energy for Farms",
+          slug: "farm-energy",
+          description: "Explore solar, wind, and other renewable energy solutions for your farm.",
+        },
+        {
+          title: "Water Conservation Techniques",
+          slug: "water-conservation",
+          description: "Learn innovative methods to manage and conserve water efficiently on your farm.",
+        },
+      ],
+    },
+    
+  ];
+
